REM-1258: Add float for fixed_amount

+++ remme/shared/forms/node_account.py
@@ -14,7 +14,7 @@ class NodeAccountGenesisForm(ProtoForm):
 
 
 class SetBetPayloadForm(ProtoForm):
-    fixed_amount = fields.IntegerField(validators=[
+    fixed_amount = fields.FloatField(validators=[
         validators.DataRequired()
     ])
     min = fields.BooleanField(validators=[
